moved models outside oif client/added components folder inside src with Gif.js,Overview.js,Navbar.js,and Login.js/fixed imports on App.js
