+++ routes/api/giphy.js
@@ -1,7 +1,7 @@
-var axios = require("axios");
+// var axios = require("axios");
 
-const giphyAPI = "JXkpKRAmyWWfdABgWzYBU3qlPkzbVD2q"
+// const giphyAPI = "JXkpKRAmyWWfdABgWzYBU3qlPkzbVD2q"
 
-axios.get("http://api.giphy.com/v1/gifs/search?q=ryan+gosling&api_key=" + giphyAPI + "&limit=1");
-gif.done(function(data) { console.log("success got data", data); });
+// axios.get("http://api.giphy.com/v1/gifs/search?q=ryan+gosling&api_key=" + giphyAPI + "&limit=1");
+// gif.done(function(data) { console.log("success got data", data); });
